+++ tests/Block/BlockTest.php
@@ -12,19 +12,12 @@ use Yiisoft\Factory\NotFoundException;
 use Yiisoft\Yii\Widgets\Block;
 use Yiisoft\Yii\Widgets\Tests\Support\TestTrait;
 
-/**
- * @psalm-suppress PropertyNotSetInConstructor
- */
+use function PHPUnit\Framework\assertSame;
+
 final class BlockTest extends TestCase
 {
     use TestTrait;
 
-    /**
-     * @throws CircularReferenceException
-     * @throws InvalidConfigException
-     * @throws NotFoundException
-     * @throws NotInstantiableException
-     */
     public function testBlock(): void
     {
         Block::widget()->id('testme')->begin();
@@ -34,12 +27,6 @@ final class BlockTest extends TestCase
         $this->assertStringContainsString('<block-testme>', $this->webView->getBlock('testme'));
     }
 
-    /**
-     * @throws CircularReferenceException
-     * @throws InvalidConfigException
-     * @throws NotFoundException
-     * @throws NotInstantiableException
-     */
     public function testBlockRenderInPlaceTrue(): void
     {
         Block::widget()->id('testme')->renderInPlace()->begin();
@@ -47,5 +34,15 @@ final class BlockTest extends TestCase
         $html = Block::end();
 
         $this->assertStringContainsString('<block-testme>', $html);
+    }
+
+    public function testZeroContent(): void
+    {
+        Block::widget()->id('test')->begin();
+        echo '0';
+        $result = Block::end();
+
+        assertSame('', $result);
+        assertSame('0', $this->webView->getBlock('test'));
     }
 }